fix: move docs markdown helpers out of the nuxt modules directory
Nuxt auto-registers every file under nuxt/modules/ as a Nuxt module, so the helper and
its test were loaded as modules and the build failed. The test file being auto-loaded
also meant the unit tests ran inside the production build.

Move both to nuxt/lib/, which neither Nuxt nor Nitro scans, leaving docs-source.ts as
the only module in that directory.
